refactor: moved more core files to util module
