+++ gptme/commands.py
@@ -7,7 +7,7 @@ from time import sleep
 from typing import Literal
 
 from . import llm
-from .export import export_chat_to_html
+from .util.export import export_chat_to_html
 from .logmanager import LogManager, prepare_messages
 from .message import (
     Message,
