+++ scripts/treeofthoughts.py
@@ -17,11 +17,15 @@ from pathlib import Path
 from typing import Literal
 
 from gptme.chat import step as _step
-from gptme.context import gather_fresh_context, get_changed_files, run_precommit_checks
 from gptme.init import init
 from gptme.logmanager import Log
 from gptme.message import Message
 from gptme.prompts import get_prompt
+from gptme.util.context import (
+    gather_fresh_context,
+    get_changed_files,
+    run_precommit_checks,
+)
 from lxml import etree
 
 # Set up logging
